+++ integration-tests/build-package-cli-tests/__snapshots/map/api_BaseFeature.js
@@ -1,0 +1,2 @@
+
+//# sourceMappingURL=BaseFeature.js.map