+++ integration-tests/build-package-cli-tests/__snapshots/map/api_layers_index.js
@@ -1,0 +1,4 @@
+import '../../model/AbstractLayer.js';
+import '../../model/layers/WMSLayerImpl.js';
+import '../../model/layers/WMTSLayerImpl.js';
+//# sourceMappingURL=index.js.map