+++ packages/build-common/src/runtimeSupport/index.test.ts
@@ -48,3 +48,27 @@ describe("generateReactHooks", function () {
         `);
     });
 });
+
+describe("generateSourceId", function () {
+    it("builds a module id for unix-like paths", async function () {
+        const code = RuntimeSupport.generateSourceInfo("my-pkg", "./src/utils/helpers.ts");
+        expect(code).toMatchInlineSnapshot(`
+          "export const sourceId = "my-pkg/src/utils/helpers";
+          export default {
+          	sourceId
+          };
+          "
+        `);
+    });
+
+    it("normalizes windows-style paths and strips the extension", async function () {
+        const code = RuntimeSupport.generateSourceInfo("pkg", ".\\lib\\file.name.with.dots.tsx");
+        expect(code).toMatchInlineSnapshot(`
+          "export const sourceId = "pkg/lib/file";
+          export default {
+          	sourceId
+          };
+          "
+        `);
+    });
+});