+++ packages/build-common/types.d.ts
@@ -4,6 +4,8 @@ export { BuildConfig } from "@open-pioneer/build-support";
 import { BuildConfig } from "@open-pioneer/build-support";
 
 export namespace RuntimeSupport {
+    export type VirtualModuleType = "app" | "react-hooks" | "source-info";
+
     /** Package name of the Open Pioneer Trails runtime library. */
     export const RUNTIME_PACKAGE_NAME: string;
 
@@ -18,13 +20,21 @@ export namespace RuntimeSupport {
      * Returns the type of the virtual module or undefined if the module id does not match anything.
      *
      * Throws an error if the module id uses the virtual prefix without a match, as that might be a mistake by the user. */
-    export function parseVirtualModule(moduleId: string): "app" | "react-hooks" | undefined;
+    export function parseVirtualModule(moduleId: string): VirtualModuleType | undefined;
 
     /** Returns the module content to implement the `open-pioneer:react-hooks` module. */
     export function generateReactHooks(
         packageName: string,
         reactIntegrationModuleId?: string
     ): string;
+
+    /**
+     * Generates the module containing the sourceId of the importing file.
+     *
+     * `packageName` is the name of the package.
+     * `relativeModulePath` is the path of the module relative to the package root.
+     */
+    export function generateSourceInfo(packageName: string, relativeModulePath: string): string;
 }
 
 /**