+++ packages/vite-plugin/test-data/codegen-source-info-edge-cases/apps/test-app/app.js
@@ -1,0 +1,7 @@
+import { logSourceId as sourceIdOutSidePackage } from "../../log";
+import { Component } from "./Component";
+
+// Ensure components are not optimized out
+console.log(Component);
+
+sourceIdOutSidePackage();