+++ packages/build-common/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog @open-pioneer/build-common
 
+## 3.2.0
+
+### Minor Changes
+
+- 19aa36f: Implement sourceId helper which provides an easy way to obtain an id for the current source file:
+
+    ```ts
+    import { sourceId } from "open-pioneer:source-info";
+
+    // If imported from foo/bar/baz.ts in package my-package: my-package/foo/bar/baz
+    console.log(sourceId);
+    ```
+
+## 3.1.2
+
+### Patch Changes
+
+- 2c45a18: Bump dependencies
+
 ## 3.1.1
 
 ### Patch Changes